requirements.txt
fastapi
uvicorn
sqlmodel
psycopg2-binary
python-dotenv
qrcode
pillow